+++ __init__.py
@@ -21,6 +21,7 @@ MODULES = [
     "core.vrm_integration",
     "core.mtoon_materials",
     "core.facial_video_preview",
+    "core.strip_remesh",
     "core.qremeshify_runtime.lib.data",
     "core.qremeshify_runtime.lib",
     "core.qremeshify_runtime.util.bisect",